+++ background.py
@@ -1,6 +1,6 @@
 from queue import Queue
 import threading
-from serialserver import data_lock, data_dict
+import shared_data
 
 def latest_data_dict(queue: Queue):
     """
@@ -11,7 +11,7 @@ def latest_data_dict(queue: Queue):
         if not queue.empty():
             data = queue.get()
             print(f"Latest data: {data}")
-            with data_lock:
-                data_dict[data[1]] = data[0]  # Assuming data is a tuple (parsed_data, parser_name)
+            with shared_data.data_lock:
+                shared_data.data_dict[data[1]] = data[0]  # Assuming data is a tuple (parsed_data, parser_name)
         else:
             threading.Event().wait(0.05)  # Wait for 0.05 seconds before checking again
