+++ shared_data.py
@@ -1,0 +1,5 @@
+import threading
+
+# Shared data and lock for thread-safe access
+data_lock = threading.Lock()
+data_dict = {}
